% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_junctions.R
\name{border_matrix}
\alias{border_matrix}
\title{Add an external border to a matrix}
\usage{
border_matrix(m)
}
\arguments{
\item{m}{The matrix to border}
}
\value{
The matrix m with a 1 column and 1 row border of 0s
}
\description{
Add an external border to a matrix
}
\examples{
m <- matrix(1, 3, 3)
border_matrix(m)
}
