% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_junctions.R
\name{find_adj_stack}
\alias{find_adj_stack}
\title{Find adjancencies by stacking shifted versions of a matrix}
\usage{
find_adj_stack(m, bordermat = NULL)
}
\arguments{
\item{m}{A binary map (matrix).}

\item{bordermat}{m, but with a 1 column and 1 row border of 0s. Generated by \link{border_matrix}}
}
\value{
A 3D array of the size of m with 8 layers. Each layer corresponds to the pixel one pixel in a given direction of the moore neighbourhood.
}
\description{
Find adjancencies by stacking shifted versions of a matrix
}
\examples{
m <- generate_test_matrix(5)
find_adj_stack(m)
}
