% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njar-utils.R
\name{generate_test_matrix}
\alias{generate_test_matrix}
\title{Generate a random binary map}
\usage{
generate_test_matrix(n = 500)
}
\arguments{
\item{n}{The size of the map}
}
\value{
A random binary map (matrix)
}
\description{
Generate a random binary map
}
\examples{
generate_test_matrix(5)
}
