% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image.R
\name{load_image}
\alias{load_image}
\title{Load an image into njar}
\usage{
load_image(p, skeletonize = TRUE, return_matrix = FALSE)
}
\arguments{
\item{p}{The path to a binary map to load.}

\item{skeletonize}{Whether to skeletonize the image using scikit-image through the reticulate package.}

\item{return_matrix}{Whether to return the image as a matrix.}
}
\value{
An \link{NJANet} object or \code{matrix} containing the image data.
}
\description{
Load an image and optionally skeletonize it.

If you have already skeltonized the image externally, use \code{\link[=load_skel]{load_skel()}}.
}
\section{Binary image requirements}{

The provided image should be in .png, .jpg, or .bmp format.

Fully white pixels should indicate places where the network is present.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
load_image("path/to/image.png")
\dontshow{\}) # examplesIf}
}
