% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image.R
\name{load_skel}
\alias{load_skel}
\title{Load a pre-skeletonized image into njar}
\usage{
load_skel(p, return_matrix = FALSE, orig_image = NULL)
}
\arguments{
\item{p}{The path to a binary map to load.}

\item{return_matrix}{Whether to return the image as a matrix.}

\item{orig_image}{(optional) The original image used to generate the skeleton.}
}
\value{
An \link{NJANet} object or \code{matrix} containing the image data.
}
\description{
Load an image that has already been skeletonized

If you have a non-skeletonized image, use \code{\link[=load_image]{load_image()}}.
}
\section{Binary image requirements}{

The provided image should be in .png, .jpg, or .bmp format.

Fully white pixels should indicate places where the network is present.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
load_skel("path/to/skeleton.png")
\dontshow{\}) # examplesIf}
}
