% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_junctions.R
\name{map_junctions}
\alias{map_junctions}
\title{Map the number of white pixels that each white pixel neighbours}
\usage{
map_junctions(m, bordermat = NULL, adj_stack = NULL)
}
\arguments{
\item{m}{A binary map (matrix).}

\item{bordermat}{m, but with a 1 column and 1 row border of 0s. Generated by \link{border_matrix}}

\item{adj_stack}{A 3D array containing the stack of adjacencies for each pixel.}
}
\value{
A matrix of the same dimensions as m, but with each cell containing the number of neighbouring white pixels (if the cell itself is white)
}
\description{
Map the number of white pixels that each white pixel neighbours
}
\examples{
m <- generate_test_matrix(5)
map_junctions(m)
}
