% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NJANet.R
\name{NJANet}
\alias{NJANet}
\title{NJA network class}
\value{
A new \link{NJANet} object.
}
\description{
The container of an NJA network.
}
\concept{R6}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{image}}{The original image used for processing.}

\item{\code{skel}}{A skeletonised version of the original image.}

\item{\code{nodes}}{The nodes of the network}

\item{\code{edges}}{The edges of the network}

\item{\code{incorrect_1nodes}}{Nodes with 1 junc and more than 1 connected edge}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{bordermat}}{skel but with a 1 column and 1 row border of 0s.}

\item{\code{juncmat}}{A map of the number of junctions that each cell has.}

\item{\code{adj_stack}}{A 3D array containing the stack of adjacencies for each pixel.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NJANet-new}{\code{NJANet$new()}}
\item \href{#method-NJANet-print}{\code{NJANet$print()}}
\item \href{#method-NJANet-clone}{\code{NJANet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NJANet-new"></a>}}
\if{latex}{\out{\hypertarget{method-NJANet-new}{}}}
\subsection{Method \code{new()}}{
Create a new NJANet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NJANet$new(
  image = NULL,
  skel = NULL,
  nodes = NULL,
  edges = NULL,
  bordermat = NULL,
  juncmat = NULL,
  adj_stack = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image}}{The original image used for processing.}

\item{\code{skel}}{A skeletonised version of the original image.}

\item{\code{nodes}}{The nodes of the network}

\item{\code{edges}}{The edges of the network}

\item{\code{bordermat}}{skel but with a 1 column and 1 row border of 0s.}

\item{\code{juncmat}}{A map of the number of junctions that each cell has.}

\item{\code{adj_stack}}{A 3D array containing the stack of adjacencies for each pixel.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NJANet-print"></a>}}
\if{latex}{\out{\hypertarget{method-NJANet-print}{}}}
\subsection{Method \code{print()}}{
Print an \link{NJANet} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NJANet$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NJANet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NJANet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NJANet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
