% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NJANode.R
\name{NJANode}
\alias{NJANode}
\title{NJA node class}
\value{
A new \link{NJANode} object.
}
\description{
A node in the NJA graph. Make a new one with \code{NJANode$new()}
}
\concept{R6}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pos}}{The position of the node.}

\item{\code{surround}}{A 3x3 array indicating the pixel context of the node.}

\item{\code{juncs}}{The number of junctions the node has.}

\item{\code{dirs}}{The directions that junctions of the node emit.}

\item{\code{uid}}{A string of the position, acting as a unique identifier.}

\item{\code{connected_edges}}{A named list of all connected edges.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NJANode-new}{\code{NJANode$new()}}
\item \href{#method-NJANode-print}{\code{NJANode$print()}}
\item \href{#method-NJANode-clone}{\code{NJANode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NJANode-new"></a>}}
\if{latex}{\out{\hypertarget{method-NJANode-new}{}}}
\subsection{Method \code{new()}}{
Create a new NJANode object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NJANode$new(pos, surround = NULL, juncs = NULL, dirs = NULL, uid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pos}}{The position of the node.}

\item{\code{surround}}{A 3x3 array indicating the pixel context of the node.}

\item{\code{juncs}}{The number of junctions the node has.}

\item{\code{dirs}}{The directions that junctions of the node emit.}

\item{\code{uid}}{A string of the position, acting as a unique identifier.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NJANode-print"></a>}}
\if{latex}{\out{\hypertarget{method-NJANode-print}{}}}
\subsection{Method \code{print()}}{
Print an \link{NJANode} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NJANode$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NJANode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NJANode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NJANode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
