% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njar-package.R
\docType{package}
\name{njar-package}
\alias{njar}
\alias{njar-package}
\title{njar: Network fitting from binary maps in R}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fwimp/njar}
  \item \url{https://fwimp.github.io/njar/}
  \item Report bugs at \url{https://github.com/fwimp/njar/issues}
}

}
\author{
\strong{Maintainer}: Francis Windram \email{francis.windram17@imperial.ac.uk} (\href{https://orcid.org/0000-0002-2129-826X}{ORCID})

}
\keyword{internal}
