% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_path.R
\name{trace_all_paths}
\alias{trace_all_paths}
\title{Trace the path between all junctions in a binary map}
\usage{
trace_all_paths(m, bordermat = NULL, juncmat = NULL)
}
\arguments{
\item{m}{A binary map (matrix).}

\item{bordermat}{m, but with a 1 column and 1 row border of 0s. Generated by \link{border_matrix}}

\item{juncmat}{A map of the number of junctions that each cell has. Generated by \link{map_junctions}}
}
\value{
A list of lists, will eventually return an NJANet object I think.
}
\description{
Trace the path between all junctions in a binary map
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
m <- matrix(FALSE, 5, 5)
m[3,] <- TRUE
m[4:5,5] <- TRUE
trace_all_paths(m)
\dontshow{\}) # examplesIf}
}
