% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_path.R
\name{trace_path}
\alias{trace_path}
\title{Trace the path between two junctions over a binary map}
\usage{
trace_path(m, startcoords, direction, bordermat = NULL, juncmat = NULL)
}
\arguments{
\item{m}{A binary map (matrix).}

\item{startcoords}{A vector of the coordinates specifying the start of the path.}

\item{direction}{The direction to take the first step in (either a number 1:9 or a direction, e.g. 2 or 'w').}

\item{bordermat}{m, but with a 1 column and 1 row border of 0s. Generated by \link{border_matrix}}

\item{juncmat}{A map of the number of junctions that each cell has. Generated by \link{map_junctions}}
}
\value{
A list containing the start and end coordinates of the path, the distance travelled, and the path taken.
}
\description{
Trace the path between two junctions over a binary map
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
m <- matrix(FALSE, 5, 5)
m[3,] <- TRUE
m[4:5,5] <- TRUE
trace_path(m, c(3,1), "e")
\dontshow{\}) # examplesIf}
}
